[Refractor] 디렉토리 수정 및 클래스 이름 변경
