+++ routers/challenge.py
@@ -6,10 +6,9 @@ from typing import List, Dict, Any
 from db.session import get_db
 from schemas.challenge import (ChallengeCreate, ChallengeJoin, ChallengeListRead, 
     ChallengeRead, ChallengeDetailRead)
-from services.challenge.createService import ChallengeCreateService
-from services.challenge.joinService import ChallengeJoinService
-from services.challenge.searchService import ChallengeSearchService
+from services.challenge.basicService import ChallengeBasicService
 from services.challenge.readService import ChallengeReadService
+from services.challenge.detailService import ChallengeDetailService
 from auth.jwt_token import get_current_user 
 
 router = APIRouter(prefix="/challenges", tags=["challenges"])
@@ -21,7 +20,7 @@ def create_challenge(
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user) 
 ):
-    ChallengeCreateService.create_challenge(db, current_user.userId, challenge_data)
+    ChallengeBasicService.create_challenge(db, current_user.userId, challenge_data)
     return {"message": "챌린지를 성공적으로 생성했습니다."}
 
 # 챌린지 참여 라우터
@@ -31,7 +30,7 @@ def join_challenge(
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user)
 ):
-    ChallengeJoinService.join_challenge(db, current_user.userId, join_data)
+    ChallengeBasicService.join_challenge(db, current_user.userId, join_data)
     return {"message": "챌린지에 성공적으로 참여했습니다."}
 
 # 챌린지 검색 라우터
@@ -40,7 +39,7 @@ def search_challenge(
     name: str,
     db: Session = Depends(get_db)
 ):
-    result = ChallengeSearchService.search_challenge(db, name)
+    result = ChallengeReadService.search_challenge(db, name)
 
     if not result:
         return JSONResponse(
@@ -71,4 +70,4 @@ def read_challenge_detail(
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user)
 ):
-    return ChallengeReadService.read_challenge_detail(db, challengeId, current_user.userId)
+    return ChallengeDetailService.read_challenge_detail(db, challengeId, current_user.userId)
